+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - completion for labeled `spa` block [#98](https://github.com/avenga/couper-vscode/pull/98)
 - completion for `configuration_backend`, `jwks_uri_backend`, `token_backend` and `userinfo_backend` attributes in `oidc` block [#98](https://github.com/avenga/couper-vscode/pull/98)
+- checks for invalid block labels [#98](https://github.com/avenga/couper-vscode/pull/98)
 
 ### Removed